requirements.txt
pillow
pytesseract